138 functionality for deleting user (#143)
* add: shadcn alertdialog

* functionality for deleting user account

* prevent deletion of test account

* fix: lint

+++ backend/src/resolvers/userResolver.ts
@@ -1,4 +1,4 @@
-import { authenticateUser } from '@/auth/utils'
+import { authenticateUser, generateToken } from '@/auth/utils'
 import { s3 } from '@/s3/s3Provider'
 import { ApolloContext } from '@/server'
 import createUpdateQuery from '@/utils/createUpdateQuery'
@@ -83,6 +83,7 @@ const UserResolver: Resolvers = {
     createUser: async (
       _: unknown,
       { name, username, password }: { name: string; username: string; password: string },
+      contextValue: ApolloContext,
     ) => {
       console.log('Creating user with name: ', name, ' username: ', username)
       try {
@@ -107,6 +108,7 @@ const UserResolver: Resolvers = {
         console.log('Created user: ', username, ' with id: ', rows[0].id)
 
         const user = rows[0] as User
+        generateToken(contextValue.res, user)
 
         return user
       } catch (error) {
@@ -121,7 +123,10 @@ const UserResolver: Resolvers = {
         throw new ApolloError('Cannot update the test user', 'DEFAULT_USER_UPDATE')
       }
       // Have to safe parse as the image is a FileUpload type and not a File type
-      const result = UpdateUserSchema.safeParse(user)
+      const result = UpdateUserSchema.safeParse({
+        ...user,
+        confirmPassword: user.password,
+      })
       let imgFile = null
 
       if (!result.success) {
@@ -133,7 +138,9 @@ const UserResolver: Resolvers = {
             throw new ApolloError('Invalid image type', 'VALIDATION_ERROR')
           }
         } else {
-          throw new ApolloError(result.error.message, 'VALIDATION_ERROR')
+          if (result.error instanceof ZodError) {
+            throw new ApolloError('Taper', 'VALIDATION_ERROR')
+          }
         }
       }
 
@@ -250,6 +257,9 @@ const UserResolver: Resolvers = {
     },
 
     deleteUser: async (_: unknown, { id }: { id: string }, contextValue: ApolloContext) => {
+      if (id === '340c0679-5f34-45e0-8189-f5929c2ebd2c') {
+        throw new ApolloError('Cannot delete the test user', 'DEFAULT_USER_DELETE')
+      }
       authenticateUser(contextValue, id)
       try {
         const query = 'DELETE FROM users WHERE id = $1 RETURNING *'
